+++ lang/cs/messages.php
@@ -8,4 +8,9 @@ return [
 
     'translated' => 'Přeloženo',
 
+    'email.hello' => 'Dobrý den!',
+    'email.regards' => 'S pozdravem',
+    'email.signature' => 'Tým :company',
+    'email.trouble' => 'Pokud máte problém s kliknutím na tlačítko, klikněte na URL níže nebo ji zkopírujte a vložte do svého webového prohlížeče.',
+
 ];